feat(sort): Implement and test insertion sort algorithm
The sort_insertion function's nested for loop's loop control variable (j) could be initialized as (i - 1), for loop's condition (j < i) could be replaced with (j >= 0), and all (i - 1 - j) could be replaced with (j). However, the reason it is implemented as is, is to be able to maintain the range of j between [0, i - 1]. Using the alternate solution would cause j to "wrap around" or "underflow" as it is a size_t variable (unsigned). This would nullify the (j >= 0) condition and violate the range [0, i - 1].